feat: Simplify hover
Just show a link to run the 'inspect' command.
